+++ test/telegram.js
@@ -880,6 +880,10 @@ describe('TelegramBot', function telegramSuite() {
     });
   });
 
+  describe.skip('#banChatSenderChat', function banChatSenderChatSuite() { });
+
+  describe.skip('#unbanChatSenderChat', function banChatSenderChatSuite() { });
+
   describe('#setChatPermissions ', function setChatPermissionsSuite() {
     it('should set chat permissions', function test() {
       const ChatPermissions = {
@@ -904,7 +908,7 @@ describe('TelegramBot', function telegramSuite() {
     });
     it('should export the group invite link', function test() {
       return bot.exportChatInviteLink(GROUPID).then(resp => {
-        assert(resp.match(/^https:\/\/t\.me\/joinchat\/.+$/i), 'is a telegram invite link');
+        assert(resp.match(/^https:\/\/t\.me\/.+$/i), 'is a telegram invite link');
       });
     });
   });
@@ -918,7 +922,7 @@ describe('TelegramBot', function telegramSuite() {
     });
     it('should create a chat invite link', function test() {
       return bot.createChatInviteLink(GROUPID).then(resp => {
-        assert(resp.invite_link.match(/^https:\/\/t\.me\/joinchat\/.+$/i), 'is a telegram invite link');
+        assert(resp.invite_link.match(/^https:\/\/t\.me\/.+$/i), 'is a telegram invite link');
         inviteLink = resp.invite_link;
       });
     });